pyproject.toml
[project]
name = "langgraph-project-based-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "amadeus>=12.0.0",
    "chromadb>=1.4.0",
    "ddgs>=9.10.0",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.5",
    "langchain-experimental>=0.4.1",
    "langchain-google-genai>=4.1.2",
    "langchain-openai>=1.1.6",
    "langchain-opentutorial>=0.0.8",
    "langgraph>=1.0.5",
    "langgraph-checkpoint-sqlite>=3.0.1",
    "matplotlib>=3.10.8",
    "pdfminer>=20191125",
    "pdfminer-six>=20251230",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
]
